Make MCFG_DEVICE_ADD and callable device types more flexible: * Allows defaulted clocks (see subtle example with vboy) * Allows additional constructors (see RS232 port in tranz330) * Allows use of device finder in place of tag in MCFG_DEVICE_ADD * Requires out-of-line destructor for devices using incomplete types * Requires XTAL or explicit u32 for clocks for devices with private types
Devices must still define the standard constructor.  When writing
additional constructors, be aware that the constructor runs before
device_add_mconfig in the context of the existing device, not the new
device.  See osborne1, zorba, tranz330, and vboy for examples of this in
use.  Compilation is a bit slower, but this is temporary while
refactoring is in progress.

Eliminated the need for MCFG_SOUND_ROUTE_EX.

Removed macros from slot option configuration - they just obfuscated
code and slowed it down with needless dynamic casts, but didn't actually
simplify it.

+++ src/mame/drivers/aim65.cpp
@@ -194,15 +194,15 @@ MACHINE_CONFIG_START(aim65_state::aim65)
 	MCFG_DEFAULT_LAYOUT(layout_aim65)
 
 	/* alpha-numeric display */
-	MCFG_DEVICE_ADD("ds1", DL1416T, 0)
+	MCFG_DEVICE_ADD("ds1", DL1416T, u32(0))
 	MCFG_DL1416_UPDATE_HANDLER(WRITE16(aim65_state, aim65_update_ds<1>))
-	MCFG_DEVICE_ADD("ds2", DL1416T, 0)
+	MCFG_DEVICE_ADD("ds2", DL1416T, u32(0))
 	MCFG_DL1416_UPDATE_HANDLER(WRITE16(aim65_state, aim65_update_ds<2>))
-	MCFG_DEVICE_ADD("ds3", DL1416T, 0)
+	MCFG_DEVICE_ADD("ds3", DL1416T, u32(0))
 	MCFG_DL1416_UPDATE_HANDLER(WRITE16(aim65_state, aim65_update_ds<3>))
-	MCFG_DEVICE_ADD("ds4", DL1416T, 0)
+	MCFG_DEVICE_ADD("ds4", DL1416T, u32(0))
 	MCFG_DL1416_UPDATE_HANDLER(WRITE16(aim65_state, aim65_update_ds<4>))
-	MCFG_DEVICE_ADD("ds5", DL1416T, 0)
+	MCFG_DEVICE_ADD("ds5", DL1416T, u32(0))
 	MCFG_DL1416_UPDATE_HANDLER(WRITE16(aim65_state, aim65_update_ds<5>))
 
 	/* Sound - wave sound only */
